Fix warnings in applications/mc/spins
- abstract_fitter.C: fix printf format mismatch — use %ld with
  static_cast<long> for boost posix_time hour/min/sec types (long long)
- base_incr_fitter.C: fix mismatched delete/new[] — use delete[]
  for array allocated with new int[]
- spinsim.h: remove two unused typedef 'parent' declarations

+++ applications/mc/spins/abstract_fitter.C
@@ -1148,9 +1148,9 @@ void AbstractFitter::print_results(const std::string filename) const
   char time_measure[256];
   boost::posix_time::time_duration time_used;
   time_used = boost::posix_time::second_clock::local_time() - start_time;
-  snprintf(time_measure,sizeof(time_measure), "\nRan %i hours, %i minutes %i seconds.",
-               time_used.hours(), time_used.minutes(),
-               time_used.seconds());
+  snprintf(time_measure,sizeof(time_measure), "\nRan %ld hours, %ld minutes %ld seconds.",
+               static_cast<long>(time_used.hours()), static_cast<long>(time_used.minutes()),
+               static_cast<long>(time_used.seconds()));
 
   mean_stream << "\n" << time_measure << std::endl;
   error_stream << "\n" << time_measure << std::endl;
